clean up value-name usage in hyper client bool options
value-name is only used to generate help message
and manpage in go-flags. The bool option ignores
value-name in help message and prints it's own default
value in the help message and manpage. So value-name is
useless for go-flags bool option.

BTW, exec's Attach option defaults to true and it
seems to be ignored anyway?

Signed-off-by: Peng Tao <bergwolf@gmail.com>