add new StorageDriver: rawblock
The driver creates raw-file-block for every images/valume.
It likes vfs driver in the operation creating an image from
the parent. Both drivers do the operation by copying, but
rawblock driver tries to copy the parent by reflink.
So the copy-on-write is supported for the rawblock driver
if the backing(host) filesystem supports copy-on-write.

If the backing filesystem is btrfs, rawblock driver
supports copy-on-write now. If the backing filesystem
is xfs, the rawblock driver will test whether it is
reflink-enabled, if true, rawblock driver supports
copy-on-write.

rawblock driver vs vfs driver:
  hypervisor handles block better than dir(9p, sharefs).
  rawblock driver supports copy-on-write

rawblock driver vs devicemapper:
  loopback-based devicemapper is too ugly.
  rawblock driver (will) use loopback only when manipulate
    the filesystem in host.

The rawblock driver is also the base feature prepared for
dax support and pages-shared across VMs.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>