use govendor instead of godep for the nested vendor issue
godep will introduce a nested-vendor issue when update a dep:

```
godep update github.com/hyperhq/runv/...
```

will include `github.com/hyperhq/runv/vendor/*` to vendor/ , and this is still an open issue for
 godep https://github.com/tools/godep/issues/498 and has last for more than 8 months.
Move to govendor could solve this.

However, official `go dep` tool is under development (which is not release and not fit for daily
usage yet) and plan to release with go 1.10 by the end of 2017. I am not sure whether we should
 change a vendor tool just for about 9 months. ref: https://github.com/golang/dep/wiki/Roadmap

Signed-off-by: Wang Xu <gnawux@gmail.com>