close write stream when finish attaching
The container logger copier expect to read EOF so that it can determine
the container output stream has ended. If we do not close stdout/stderr
streams, the logger copier will continue waiting for a new liner to stop
reading. When container stops and closes the logger, container logs are
lost forever.

Signed-off-by: Peng Tao <bergwolf@gmail.com>