add missing mapping to interface id
Without mapping this field, interfaces will be persisted with a 0 id
resulting in failed load on startup. This results in the entire pod and
its containers not being loaded, and hyperd is not aware of them.
However when a new pod with a container of the same name is created, the
create will fail due to the name conflict with the previously existing
container.