persist.go: saveInterface: loadInterface: separate inf.descript.Id and interface id to fix restore issue
Current saveInterface and loadInterface handle the inf.descript.Id and
interface id as same thing that is not right.
inf.descript.Id will be reset in addInterface in runv.

So add this patch to separte inf.descript.Id and interface id to handle
the issue.

And change var name to iId to make the code clear in LoadXPod and savePod.

Signed-off-by: Hui Zhu <teawater@hyper.sh>