buffer: Handle the control operations (ContainerCreate) asynchronously
Currently, hyperd handles all the control operations synchronously.
K8S will wait the result long time if there are a lot of operations
in same time.

This patch put control operations (just support ContainerCreate now)
to buffer and put back the result at once.
The buffer will handle the result in goroutins pool (the size can be
config by BufferGoroutinesMax and BufferChannelSize).

When open buffer function, hyperd can pass most of test in hyper_test.go
but not TestRenameContainer and TestSendContainerSignal.
Because after container created, the id of container will be changed to
the id that get from docker.

Signed-off-by: Hui Zhu <teawater@hyper.sh>